fix(home/AGN-710): move dynamic ssr:false into client wrapper (Next 15 RSC)
